# nodejs-server-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "User related user stories API."
  version: "0.1.9"
  title: "GreenHero API"
...
paths:
  /User:
    post:
      tags:
      - "User"
      summary: "Create User"
      operationId: "createUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "To create a new user"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "201":
          description: "successful operation"
      x-swagger-router-controller: "User"
  /User/login:
    get:
      tags:
      - "User"
      summary: "Logs user into the system"
      operationId: "loginUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "userName"
        in: "query"
        description: "The user name for login"
        required: true
        type: "string"
      - name: "userPassword"
        in: "query"
        description: "The password for login in clear text"
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
        "400":
          description: "Invalid username/password supplied"
      x-swagger-router-controller: "User"
  /User/{userName}:
    get:
      tags:
      - "User"
      summary: "Get user by user name"
      description: ""
      operationId: "getUserByUserName"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "userName"
        in: "path"
        description: "The name that needs to be fetched. Use user1 for testing. "
        required: true
        type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/User"
        "400":
          description: "Invalid username supplied"
        "404":
          description: "User not found"
      x-swagger-router-controller: "User"
    put:
      tags:
      - "User"
      summary: "Modify user"
      description: "To modify a a username/Password "
      operationId: "modifyUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "userName"
        in: "path"
        description: "The username of the user that needs to be modified"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/User"
      responses:
        "200":
          description: "updated successfully"
        "204":
          description: "No content "
        "404":
          description: "UserName not found "
      x-swagger-router-controller: "User"
    delete:
      tags:
      - "User"
      summary: "Delete user"
      description: "To delete a user entry"
      operationId: "deleteUserByUserName"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "userName"
        in: "path"
        description: "The username of the user that needs to be deleted"
        required: true
        type: "string"
      responses:
        "200":
          description: "Deleted sucessfully"
        "404":
          description: "User not found"
      x-swagger-router-controller: "User"
...
  /Team:
    get:
      tags:
      - "Team"
      summary: "Get teams"
      description: "See the available teams"
      operationId: "getAllTeams"
      produces:
      - "application/xml"
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Team"
        "404":
          description: "No team found"
      x-swagger-router-controller: "Team"
    post:
      tags:
      - "Team"
      summary: "Create a new team"
      operationId: "createTeam"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "To create a new team"
        required: true
        schema:
          $ref: "#/definitions/Team"
      responses:
        "201":
          description: "successful operation"
      x-swagger-router-controller: "Team"
    put:
      tags:
      - "Team"
      summary: "Modify team"
      description: "To modify a team"
      operationId: "modifyTeam"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Updated team object"
        required: true
        schema:
          $ref: "#/definitions/Team"
      responses:
        "200":
          description: "updated successfully"
        "204":
          description: "No content "
        "404":
          description: "team not found "
      x-swagger-router-controller: "Team"
  /Team/{teamName}:
    delete:
      tags:
      - "Team"
      summary: "Delete team"
      description: "To delete a team"
      operationId: "deleteTeamByTeamName"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "teamName"
        in: "path"
        description: "The name of the team that needs to be deleted"
        required: true
        type: "string"
      responses:
        "200":
          description: "Deleted sucessfully"
        "404":
          description: "team not found"
      x-swagger-router-controller: "Team"
...
definitions:
  User:
    type: "object"
    properties:
      _id:
        type: "string"
      UserName:
        type: "string"
      Password:
        type: "string"
      email:
        type: "string"
        format: "email"
      personalInfo:
        type: "string"
      typeOfUser:
        type: "string"
      avatarUrl:
        type: "string"
      abilityPoints:
        type: "integer"
      consumption:
        type: "string"
      allowance:
        $ref: "#/definitions/User_allowance"
    xml:
      name: "User"
    example:
      personalInfo: "personalInfo"
      typeOfUser: "typeOfUser"
      abilityPoints: 0
      UserName: "UserName"
      avatarUrl: "avatarUrl"
      consumption: "consumption"
      allowance:
        "01_2020": 6.02745618307040320615897144307382404804229736328125
        "10_2020": 4.1456080298839363962315474054776132106781005859375
        "07_2020": 9.301444243932575517419536481611430644989013671875
        "09_2020": 2.027123023002321833274663731572218239307403564453125
        "02_2020": 1.46581298050294517310021547018550336360931396484375
        "06_2020": 7.061401241503109105224211816675961017608642578125
        "05_2020": 2.3021358869347654518833223846741020679473876953125
        "11_2020": 7.3862819483858839220147274318151175975799560546875
        "03_2020": 5.962133916683182377482808078639209270477294921875
        "04_2020": 5.63737665663332876420099637471139430999755859375
        "08_2020": 3.61607674925191080461672754609026014804840087890625
        "12_2020": 1.231513536777255612975068288506008684635162353515625
      _id: "_id"
      email: "email"
      Password: "Password"
...
  Team:
    type: "object"
    properties:
      _id:
        type: "string"
      teamName:
        type: "string"
      avatar:
        type: "string"
      event_id:
        type: "string"
      teamLeader:
        type: "string"
      teamMembers:
        type: "array"
        xml:
          name: "teamMembers"
          wrapped: true
        items:
          type: "string"
    xml:
      name: "Team"
    example:
      teamName: "teamName"
      event_id: "event_id"
      _id: "_id"
      avatar: "avatar"
      teamLeader: "teamLeader"
      teamMembers:
      - "teamMembers"
      - "teamMembers"
...
  User_allowance:
    type: "object"
    properties:
      "01_2020":
        type: "number"
      "02_2020":
        type: "number"
      "03_2020":
        type: "number"
      "04_2020":
        type: "number"
      "05_2020":
        type: "number"
      "06_2020":
        type: "number"
      "07_2020":
        type: "number"
      "08_2020":
        type: "number"
      "09_2020":
        type: "number"
      "10_2020":
        type: "number"
      "11_2020":
        type: "number"
      "12_2020":
        type: "number"
    example:
      "01_2020": 6.02745618307040320615897144307382404804229736328125
      "10_2020": 4.1456080298839363962315474054776132106781005859375
      "07_2020": 9.301444243932575517419536481611430644989013671875
      "09_2020": 2.027123023002321833274663731572218239307403564453125
      "02_2020": 1.46581298050294517310021547018550336360931396484375
      "06_2020": 7.061401241503109105224211816675961017608642578125
      "05_2020": 2.3021358869347654518833223846741020679473876953125
      "11_2020": 7.3862819483858839220147274318151175975799560546875
      "03_2020": 5.962133916683182377482808078639209270477294921875
      "04_2020": 5.63737665663332876420099637471139430999755859375
      "08_2020": 3.61607674925191080461672754609026014804840087890625
      "12_2020": 1.231513536777255612975068288506008684635162353515625