# nodejs-server-server/controllers/User.js
'use strict';

var utils = require('../utils/writer.js');
var User = require('../service/UserService');

module.exports.createUser = function createUser (req, res, next) {
  var body = req.swagger.params['body'].value;
  User.createUser(body)
    .then(function (response) {
      utils.writeJson(res, response);
    })
    .catch(function (response) {
      utils.writeJson(res, response);
    });
};

module.exports.deleteUserByUserName = function deleteUserByUserName (req, res, next) {
  var userName = req.swagger.params['userName'].value;
  User.deleteUserByUserName(userName)
    .then(function (response) {
      utils.writeJson(res, response);
    })
    .catch(function (response) {
      utils.writeJson(res, response);
    });
};

module.exports.getUserByUserName = function getUserByUserName (req, res, next) {
  var userName = req.swagger.params['userName'].value;
  User.getUserByUserName(userName)
    .then(function (response) {
      utils.writeJson(res, response);
    })
    .catch(function (response) {
      utils.writeJson(res, response);
    });
};

module.exports.loginUser = function loginUser (req, res, next) {
  var userName = req.swagger.params['userName'].value;
  var userPassword = req.swagger.params['userPassword'].value;
  User.loginUser(userName,userPassword)
    .then(function (response) {
      utils.writeJson(res, response);
    })
    .catch(function (response) {
      utils.writeJson(res, response);
    });
};

module.exports.modifyUser = function modifyUser (req, res, next) {
  var userName = req.swagger.params['userName'].value;
  var body = req.swagger.params['body'].value;
  User.modifyUser(userName,body)
    .then(function (response) {
      utils.writeJson(res, response);
    })
    .catch(function (response) {
      utils.writeJson(res, response);
    });
};